pyproject.toml
[project]
name = "mkdocs_autoref_plugin"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Fabrice Normandin", email = "normandf@mila.quebec" }]
requires-python = ">=3.10"
dependencies = [
    "mkdocs-autorefs[python]>=1.2.0",
    "mkdocstrings[python]>=0.27.0",
]

[project.entry-points."mkdocs.plugins"]
custom_autoref_plugin = "mkdocs_autoref_plugin.autoref_plugin:CustomAutoRefPlugin"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "uv-dynamic-versioning>=0.2.0",
    "mkdocs-material>=9.5.44",
]

[tool.pytest.ini_options]
testpaths = ["mkdocs_autoref_plugin", "tests"]
norecursedirs = [".venv"]
addopts = ["--doctest-modules"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["tool_name"]

[tool.uv]
managed = true

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"
